Get functions defined in target package

Helper functions internally by used internally by get_test_coverage().

get_pkg_functions(pkg = ".")

get_pkg_exports(pkg = ".")

get_pkg_S3methods(pkg = ".")

get_pkg_tested_functions(pkg, from_tags, from_desc)

get_pkg_testignore(pkg)

Arguments

pkg
Character scalar. Path to target package.

from_tags
Logical scalar. Checks the files if your test directory for testthis tags. Specifically, if you have the comment #* @testing myfunction in any of your test files, myfunction will be marked as tested.

from_desc
Logical scalar. Checks the desc argument test_that(...) of the tests in your test directory for functions names. E.g. if you have a testfile that contains test_that("myfunction works", {...}), myfunction will be marked as tested.

Value

get_pkg_functions() returns a character vector of all functions defined in pkg. get_pkg_exports() returns a character vector of functions exported from pkgs NAMESPACE. get_pkg_S3methods() returns a character vector of all S3 methods exported from pkgs NAMESPACE. get_pkg_tested_functions() returns a character vector of all functions for which unit tests exist. get_pkg_testignore() returns a character vector of all functions listed in tests/testthat/_testthisignore.