Testthis Taglist

This is a simple s3 constructor that does not yet perform any checks. It's only used by get_taglist

taglist(dat)

Arguments

dat
a list

Details

Possible tags:

testfile path_to_file ... Put this in a file to manually set the name of an associated test file (relevant for test_this(), open_testfile(), etc...)

testing function_name ... put this in a testfile to mark the current file includes tests for a function