Use a tester function

Quickly create an R script that contains a function for running all tests in a predefined directory. This function powers the make_tester option of use_test_subdir() and you will likely not need to run it manually.

use_tester(
  path,
  ignore = TRUE,
  tester_path = file.path("R", "testthis-testers.R")
)

Arguments

path
character scalar. Name of the subdirectory of tests/testthat/ for which to create a tester function.

ignore
logical scalar. Add tester_path to .Rbuildignore?

tester_path
logical scalar. Path to the R script file in which to store the tester functions

Value

TRUE on success (invisibly).

See also